move common pipelines to common place
